Home commit

+++ TEMPLATES/run_function/run_hough.py
@@ -17,7 +17,7 @@ def run_hough(template):
         reader = csv.reader(f)
         for row in reader:
             joints = np.append(joints, np.array(row))
-    joints = np.array(np.split(joints, reader.line_num))
+    joints = np.array(np.split(joints, reader.line_num), dtype=np.float64)
     nodes = readJoints(joints)
     # nodes = houghAnalysis(nodes);
     # fprintf('Real spacing - Hough frame : %f\n', mean((nodes.real_spacing_hough)))
